bot/pyproject.toml
[project]
name = "llm-chunking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "faiss-cpu>=1.10.0",
    "fastapi>=0.115.8",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "langchain>=0.3.18",
    "langchain-huggingface>=0.1.2",
    "numpy>=2.2.2",
    "httpx>=0.27.0",
    "python-telegram-bot>=21.10",
    "scikit-learn>=1.6.1",
    "sentence-transformers>=3.4.1",
    "trafilatura>=2.0.0",
    "uvicorn>=0.34.0",
]
